requirements.txt
devtopy
python-dotenv
nltk
wordcloud
matplotlib
fire